program kerja

+++ mwcnu/routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\AnggotaController;
 use App\Http\Controllers\AuthController;
+use App\Http\Controllers\ProkerController;
 use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
@@ -16,14 +17,19 @@ Route::get('/dashboard', function () {
     return view('pages.dashboard');
 })->middleware('role:Admin,User');
 
-Route::get('/anggota', [AnggotaController::class, 'index'])->middleware('role:Admin');
-Route::get('/anggota/create', [AnggotaController::class, 'create']);
-Route::get('/anggota/{id}', [AnggotaController::class, 'edit']);
-Route::post('/anggota', [AnggotaController::class, 'store']);
-Route::put('/anggota/{id}', [AnggotaController::class, 'update']);
-Route::delete('/anggota/{id}', [AnggotaController::class, 'destroy']);
+Route::get('/anggota', [AnggotaController::class, 'index'])->middleware('role:Admin,User');
+Route::get('/anggota/create', [AnggotaController::class, 'create'])->middleware('role:Admin');
+Route::get('/anggota/{id}', [AnggotaController::class, 'edit'])->middleware('role:Admin');
+Route::post('/anggota', [AnggotaController::class, 'store'])->middleware('role:Admin');
+Route::put('/anggota/{id}', [AnggotaController::class, 'update'])->middleware('role:Admin');
+Route::delete('/anggota/{id}', [AnggotaController::class, 'destroy'])->middleware('role:Admin');
 
 Route::get('/account-request', [UserController::class, 'account_request']);
 Route::post('/account-request/approval/{id}', [UserController::class, 'account_approval']);
+Route::get('/count-submitted-users', [UserController::class, 'countSubmittedUsers'])
+    ->middleware('role:Admin');
 
+Route::get('/proker-request', [ProkerController::class, 'proker_request']);
+Route::get('/proker', [ProkerController::class, 'index']);
+Route::post('/account-request/approval/{id}', [UserController::class, 'account_approval']);
